add bypass email verification

+++ routes/auth/signup.js
@@ -44,11 +44,17 @@ router.post('/api/users/signup', [
 ],
 validateRequest,
 async (req, res) => {
+    const byPassEmailVerification = true;
+
     const { username, email, password } = req.body;
     
-    user = new User({username, email, password});
+    user = new User({username, email, password, active: byPassEmailVerification});
     
     await user.save();
+
+    if(byPassEmailVerification) {
+        return res.status(200).json({byPassEmailVerification});
+    }
 
     await SecretCode.deleteOne({username});
     const code = new SecretCode({
@@ -84,7 +90,7 @@ async (req, res) => {
             <h1>Hello</h1>
             <p>Thanks for registering with us</p>
             <p>Please click the link below to verify your account. Code is valid only for 30 mins </p>
-            <a href = http://${process.env.REACT_APP_URL_NAKED}/auth/verify?code=${code.code}&email=${base64email}>Verify your account</a>
+            <a href = ${process.env.REACT_APP_URL_NAKED}/auth/verify?code=${code.code}&email=${base64email}>Verify your account</a>
         `
     }
